+++ src/ui/mod.rs
@@ -4,5 +4,6 @@ pub mod draw;
 pub mod errors;
 pub mod help;
 pub mod priorities;
+pub mod sprints;
 mod table;
 pub mod transitions;